Move `MetricsResource`
